code plus consistant + cleaner
