+++ src/client/middlewares/socketIO.js
@@ -3,12 +3,22 @@ import { RTCConnection } from '../actions/connexion';
 export default socket => ({ dispatch, getState }) => {
   if (socket) {
     socket.on('action', dispatch);
-    socket.on('game', (newPeer) => {
-      if (newPeer.id !== getState().peer.id) {
-        const conn = getState().peer.connect(newPeer.id);
-        conn.on('open', () => {
-          dispatch(RTCConnection(conn));
-        });
+    socket.on('/game', (data) => {
+      const { path } = data;
+      const state = getState();
+      switch (path) {
+        case '/join': {
+          console.log('peer received --', data);
+          if (data.id !== state.peer.id) {
+            const conn = state.peer.connect(data.id);
+            conn.on('open', () => {
+              dispatch(RTCConnection(conn));
+            });
+          }
+          break;
+        }
+        default:
+          break;
       }
     });
   }
